PATCH: GH #17081: Workaround glibc bug with LC_MESSAGES
Please see the ticket for a full explanation.  This bug has been
submitted to glibc, without any real action forthcoming.

This invalidates the message cache each time the locale of LC_MESSAGES
is changed, as glibc should be doing this when uselocale changes that,
but glibc fails to do so.